refactor: 스타일 헬퍼 함수를 top-level로 분리
- CellStyleBuilder.Companion의 fontColor(), backgroundColor()를 StyleFactories.kt로 이동
- import 경로 단순화: CellStyleBuilder.Companion.fontColor → fontColor
- README, 가이드 문서 업데이트
