+++ core/src/main/kotlin/io/clroot/excel/core/dsl/CellStyleBuilder.kt
@@ -113,23 +113,4 @@ class CellStyleBuilder {
             numberFormat = numberFormat,
         )
 
-    companion object {
-        /**
-         * Creates a CellStyle with only fontColor set.
-         * Convenience function for conditional styles.
-         *
-         * @param color the font color to apply
-         * @return a CellStyle with only fontColor set
-         */
-        fun fontColor(color: Color): CellStyle = CellStyle(fontColor = color)
-
-        /**
-         * Creates a CellStyle with only backgroundColor set.
-         * Convenience function for conditional styles.
-         *
-         * @param color the background color to apply
-         * @return a CellStyle with only backgroundColor set
-         */
-        fun backgroundColor(color: Color): CellStyle = CellStyle(backgroundColor = color)
-    }
 }
